% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{suma}
\alias{suma}
\title{Funcion para sumar dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x}{numero (entero o doble)}

\item{y}{numero (entero o doble)}
}
\value{
El resultado de la suma 'x' + 'y'
}
\description{
Funcion para sumar dos numeros
}
